Replace hooks with CS version, part 1

+++ ChatTwo/GameFunctions/Chat.cs
@@ -626,9 +626,9 @@ internal sealed unsafe class Chat : IDisposable
         EurekaContextMenuTellHook!.Original(param1, playerName, worldName, world, id, param6);
     }
 
-    internal ulong GetContentIdForEntry(uint index)
+    internal ulong GetContentIdForEntry(int index)
     {
-        return Framework.Instance()->GetUiModule()->GetRaptureLogModule()->GetContentIdForLogMessage((int) index);
+        return Framework.Instance()->GetUiModule()->GetRaptureLogModule()->GetContentIdForLogMessage(index);
     }
 
     internal void SetChannel(InputChannel channel, string? tellTarget = null)
